requirements.txt
# Istall for wsgi
django==4.1.1
# Istall for check code
black==22.8.0
flake8==5.0.4
pyupgrade==2.38.2
pre-commit==2.20.0
# Install for fake greetings
faker==15.0.0
# Install for query
webargs==8.2.0

# Install for .env file
django-environ==0.9.0

# Istall for postgres
psycopg2-binary==2.9.5

# Install for form visualisation
django-crispy-forms==1.14.0

# Install for hide path of user avatar
python-ulid==1.1.0

# Istall to upload media
pillow==9.3.0

djangorestframework==3.14.0